+++ test/strongbox/core_test.clj
@@ -1010,7 +1010,7 @@
                        {:get (fn [req] {:status 500 :host "raw.githubusercontent.com" :reason-phrase "500 Server Error"})}}
 
           expected ["downloading 'short' catalogue"
-                    "failed to download file 'https://raw.githubusercontent.com/ogri-la/strongbox-catalogue/master/short-catalogue.json': 500 Server Error (HTTP 500)"]]
+                    "failed to fetch 'https://raw.githubusercontent.com/ogri-la/strongbox-catalogue/master/short-catalogue.json': 500 Server Error (HTTP 500)"]]
       (with-fake-routes-in-isolation fake-routes
         (with-running-app
           (is (= expected (logging/buffered-log